requirements.txt
aiohappyeyeballs
aiohttp
aiosignal
async-timeout
attrs
blinker
click
discord.py
frozenlist
idna
itsdangerous
Jinja2
jokeapi
MarkupSafe
multidict
propcache
python-dotenv
python-weather
simplejson
typing_extensions
urllib3
yarl
yt-dlp
apscheduler
ffmpeg-python
pytz
